# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/users": {
      "get": {
        "operationId": "UsersController_findAll_v1",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Listar todos os usuários com paginação",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UsersController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "409": {
            "description": "E-mail já está em uso."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Criar novo usuário (restrito a administradores)",
        "tags": [
          "Users"
        ]
      }
    },
    "/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findById_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado com sucesso."
          },
          "404": {
            "description": "Usuário não encontrado."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Buscar usuário por ID",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UsersController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "404": {
            "description": "Usuário não encontrado."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Atualizar usuário existente (apenas o próprio usuário)",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_delete_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Usuário removido com sucesso."
          },
          "404": {
            "description": "Usuário não encontrado."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Remover usuário (apenas o próprio usuário)",
        "tags": [
          "Users"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    }
  },
  "info": {
    "title": "User Management API",
    "description": "API para gerenciamento de usuários com cache Redis",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "Users",
      "description": "Operações relacionadas a usuários"
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      },
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "avatar": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MEMBER"
            ],
            "default": "MEMBER"
          },
          "password": {
            "type": "string",
            "minLength": 6
          }
        },
        "required": [
          "name",
          "email"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "avatar": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MEMBER"
            ]
          },
          "password": {
            "type": "string",
            "minLength": 6
          }
        }
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "securepassword123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      }
    }
  }
}